requirements.txt
astroid==3.3.9
black==25.1.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
dill==0.4.0
filelock==3.18.0
flake8==7.2.0
Flask==3.1.0
fonttools==4.57.0
fsspec==2025.3.2
huggingface-hub==0.30.2
idna==3.10
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mccabe==0.7.0
mpmath==1.3.0
mypy==1.15.0
mypy_extensions==1.1.0
networkx==3.4.2
numpy==1.26.4
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
pycodestyle==2.13.0
pyflakes==3.3.2
pylint==3.3.6
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
six==1.17.0
sympy==1.14.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
